% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states51}
\alias{states51}
\title{Longitude and latitude for 48 states + DC, for maps.}
\format{
A tibble with 13,694 rows and 8 variables:
\describe{
  \item{long}{longitude}
  \item{lat}{latitude}
  \item{order}{sequence number of each record - keep the file sorted on this}
  \item{hole}{}
  \item{piece}{}
  \item{id}{state name, lower case}
  \item{group}{not sure}
  \item{stabbr}{state abbreviation (postal)}
}
}
\source{
\url{https://github.com/wmurphyrd/fiftystater/}
}
\usage{
states51
}
\description{
A tibble containing the longitude and latitude for drawing U.S. maps,
with the District of Columbia, and with Alaska and Hawaii as an inset.
}
\keyword{datasets}
