% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.r
\name{theme_map}
\alias{theme_map}
\title{Theme for maps.}
\usage{
theme_map(base_size = 9, base_family = "")
}
\arguments{
\item{base_size}{A number.}

\item{base_family}{Not needed.}
}
\value{
The theme.
}
\description{
Theme for maps.
}
\examples{
library(tidyverse)
set.seed(1234)
tibble(stabbr=c(state.abb, "DC")) \%>\%
  mutate(mgroup=sample(c("A", "B", "C", "D"), 51, replace=TRUE)) \%>\%
  right_join(states51, by="stabbr") \%>\%
  ggplot(aes(x = long, y = lat, group = group, fill = mgroup)) +
    geom_polygon(color = "gray90", size = 0.1) +
    coord_map(projection = "albers", lat0 = 39, lat1 = 45) +
    scale_fill_brewer(type="qual") +
    theme_map()
}
